+++ android-agent/src/main/java/io/opentelemetry/android/OpenTelemetryRumBuilder.java
@@ -16,18 +16,9 @@ import io.opentelemetry.android.config.OtelRumConfig;
 import io.opentelemetry.android.features.diskbuffering.DiskBufferingConfiguration;
 import io.opentelemetry.android.features.diskbuffering.SignalFromDiskExporter;
 import io.opentelemetry.android.features.diskbuffering.scheduler.ExportScheduleHandler;
+import io.opentelemetry.android.features.network.CurrentNetworkProvider;
+import io.opentelemetry.android.features.network.NetworkAttributesSpanAppender;
 import io.opentelemetry.android.instrumentation.AndroidInstrumentation;
-import io.opentelemetry.android.instrumentation.activity.VisibleScreenTracker;
-import io.opentelemetry.android.instrumentation.anr.AnrDetector;
-import io.opentelemetry.android.instrumentation.anr.AnrDetectorBuilder;
-import io.opentelemetry.android.instrumentation.crash.CrashReporter;
-import io.opentelemetry.android.instrumentation.crash.CrashReporterBuilder;
-import io.opentelemetry.android.instrumentation.network.CurrentNetworkProvider;
-import io.opentelemetry.android.instrumentation.network.NetworkAttributesSpanAppender;
-import io.opentelemetry.android.instrumentation.network.NetworkChangeMonitor;
-import io.opentelemetry.android.instrumentation.slowrendering.SlowRenderingDetector;
-import io.opentelemetry.android.instrumentation.startup.InitializationEvents;
-import io.opentelemetry.android.instrumentation.startup.SdkInitializationEvents;
 import io.opentelemetry.android.internal.features.persistence.DiskManager;
 import io.opentelemetry.android.internal.features.persistence.SimpleTemporaryFileProvider;
 import io.opentelemetry.android.internal.processors.GlobalAttributesLogRecordAppender;
@@ -398,16 +389,6 @@ public final class OpenTelemetryRumBuilder {
                                 networkAttributesSpanAppender);
                     });
             initializationEvents.currentNetworkProviderInitialized();
-        }
-
-        // Add network change monitor if enabled (default = = true)
-        if (config.isNetworkChangeMonitoringEnabled()) {
-            addInstrumentation(
-                    app -> {
-                        NetworkChangeMonitor.create(getOrCreateCurrentNetworkProvider())
-                                .installOn(app);
-                        initializationEvents.networkMonitorInitialized();
-                    });
         }
 
         // Add span processor that appends screen attribute(s)
