+++ instrumentation/network/build.gradle.kts
@@ -21,6 +21,7 @@ android {
 dependencies {
     api(platform(libs.opentelemetry.platform))
     api(libs.opentelemetry.api)
+    api(project(":android-agent"))
     api(project(":common"))
     api(project(":instrumentation:common-api"))
     implementation(libs.androidx.core)
