+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkChangeAttributesExtractor.java
@@ -8,6 +8,9 @@ package io.opentelemetry.android.instrumentation.network;
 import static io.opentelemetry.api.common.AttributeKey.stringKey;
 import static io.opentelemetry.semconv.SemanticAttributes.NETWORK_CONNECTION_TYPE;
 
+import io.opentelemetry.android.features.network.CurrentNetwork;
+import io.opentelemetry.android.features.network.CurrentNetworkAttributesExtractor;
+import io.opentelemetry.android.features.network.NetworkState;
 import io.opentelemetry.api.common.AttributeKey;
 import io.opentelemetry.api.common.AttributesBuilder;
 import io.opentelemetry.context.Context;
