+++ instrumentation/network/src/main/java/io/opentelemetry/android/instrumentation/network/NetworkChangeMonitor.java
@@ -5,7 +5,15 @@
 
 package io.opentelemetry.android.instrumentation.network;
 
+import android.app.Application;
+import androidx.annotation.NonNull;
+import io.opentelemetry.android.OpenTelemetryRum;
+import io.opentelemetry.android.features.network.CurrentNetwork;
+import io.opentelemetry.android.features.network.CurrentNetworkProvider;
+import io.opentelemetry.android.instrumentation.AndroidInstrumentation;
 import io.opentelemetry.android.instrumentation.common.InstrumentedApplication;
+import io.opentelemetry.android.internal.services.ServiceManager;
+import io.opentelemetry.android.internal.services.applifecycle.AppLifecycleService;
 import io.opentelemetry.api.OpenTelemetry;
 import io.opentelemetry.instrumentation.api.instrumenter.AttributesExtractor;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
@@ -18,7 +26,7 @@ import java.util.List;
  * <p>This class is internal and not for public use. Its APIs are unstable and can change at any
  * time.
  */
-public final class NetworkChangeMonitor {
+public final class NetworkChangeMonitor implements AndroidInstrumentation {
 
     public static NetworkChangeMonitor create(CurrentNetworkProvider currentNetworkProvider) {
         return builder(currentNetworkProvider).build();
@@ -41,13 +49,7 @@ public final class NetworkChangeMonitor {
      * Installs the network change monitoring instrumentation on the given {@link
      * InstrumentedApplication}.
      */
-    public void installOn(InstrumentedApplication instrumentedApplication) {
-        NetworkApplicationListener networkApplicationListener =
-                new NetworkApplicationListener(currentNetworkProvider);
-        networkApplicationListener.startMonitoring(
-                buildInstrumenter(instrumentedApplication.getOpenTelemetrySdk()));
-        instrumentedApplication.registerApplicationStateListener(networkApplicationListener);
-    }
+    public void installOn(InstrumentedApplication instrumentedApplication) {}
 
     private Instrumenter<CurrentNetwork, Void> buildInstrumenter(OpenTelemetry openTelemetry) {
         return Instrumenter.<CurrentNetwork, Void>builder(
@@ -55,5 +57,17 @@ public final class NetworkChangeMonitor {
                 .addAttributesExtractor(new NetworkChangeAttributesExtractor())
                 .addAttributesExtractors(additionalExtractors)
                 .buildInstrumenter();
+    }
+
+    @Override
+    public void apply(
+            @NonNull Application application, @NonNull OpenTelemetryRum openTelemetryRum) {
+        NetworkApplicationListener networkApplicationListener =
+                new NetworkApplicationListener(currentNetworkProvider);
+        networkApplicationListener.startMonitoring(
+                buildInstrumenter(openTelemetryRum.getOpenTelemetry()));
+        ServiceManager.get()
+                .getService(AppLifecycleService.class)
+                .registerListener(networkApplicationListener);
     }
 }
